move index.{ID,IDRange} methods to separate files
